Investigate rsd population
Changes made to scripts to better reflect the data.

Logic is being processed as expected. Math is calculating as expected. Amount of dace just hits its threshold on the first year and because the system is inclined to be good for the Dace, they keep growing, but their population is being clamped to some max value. Therefore it appears that the amount is not changing or "stuck". This is not the case.

Solution: Red Dace and Insect population values need to be tuned in order for it to "look right". Must check up on data for accurate population values.